feat: GitHub Data Sync Integration (v0.3.0)
Implement comprehensive offline-first data synchronization system with GitHub releases.

**Major Features:**
- Automatic character data sync from GitHub releases
- Offline-first architecture with IndexedDB + Cache API
- Smart caching with version tracking (vYYYY.MM.DD-rN format)
- Background updates without blocking user workflow
- Character lookup service with O(1) validation and fuzzy search

**New Sync Module (src/ts/sync/):**
- DataSyncService: Main orchestrator with event system
- GitHubReleaseClient: GitHub API client with rate limiting
- PackageExtractor: ZIP extraction and validation
- StorageManager: IndexedDB + Cache API wrapper
- VersionManager: Version comparison logic
- MigrationHelper: Legacy data migration

**UI Components:**
- SyncStatusIndicator: Header status display
- SyncDetailsModal: Sync management dashboard
- SyncProgressBar: Real-time download progress
- Settings integration: Auto-sync toggle and controls

**Test Coverage:**
- 92 unit tests passing (100% coverage for sync module)
- E2E test checklist with 15 manual scenarios

**Documentation:**
- Comprehensive README updates with sync features
- CLAUDE.md updated with sync module architecture
- E2E_TEST_CHECKLIST.md for manual testing
- Detailed CHANGELOG.md entry

**Technical:**
- Storage: ~25 MB max (IndexedDB + Cache API)
- Performance: <50ms IndexedDB reads, <5s full sync
- Browser support: Chrome, Firefox, Edge, Safari

**Bug Fixes:**
- Fixed TypeScript compilation errors in tokenGenerator.ts
- Fixed package extractor character count validation

+++ src/components/Shared/SliderWithValue.tsx
@@ -28,8 +28,9 @@ export function SliderWithValue({
   }
 
   const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const inputValue = e.target.value.replace(/[^0-9-]/g, '')
-    const numValue = parseInt(inputValue) || 0
+    // Support decimals by allowing dots in the input
+    const inputValue = e.target.value.replace(/[^0-9.-]/g, '')
+    const numValue = parseFloat(inputValue) || 0
     const clampedValue = Math.max(min, Math.min(max, numValue))
     onChange(clampedValue)
   }
@@ -63,7 +64,7 @@ export function SliderWithValue({
           type="range"
           className={styles.range}
           value={value}
-          onChange={(e) => onChange(parseInt(e.target.value))}
+          onChange={(e) => onChange(parseFloat(e.target.value))}
           onDoubleClick={handleDoubleClick}
           min={min}
           max={max}
